fetch: allow to don't stop on old workflow run
It is useful, when we want to collect a lot of metainformation and it is
not possible at one script invocation. So, we'll call the script several
times with the --nostop option.

I want to use it to collect metainfo for all branches (without logs) to
calculate a time we spent in jobs at sum.