gather_data: fix time_queued calculating
Before this patch, there was an error occured on the `gathering_data.py`
script:

```
ValueError: Invalid isoformat string: '\ufeff2024-03-01T06:02:0+00:00'
```

This error occures, if there is a BOM at the file first byte. Before
this patch, we calculated `time_queued` as first 20 symbols of the first
log line. This patch fixes the problem with reqex search the ISO format
date and time.